refactoring kensetsu indexer types
