+++ src/modules/vault/types.ts
@@ -11,34 +11,6 @@ export type VaultEvent = {
   type: VaultEventType;
 };
 
-export type IndexerVaultEvent = {
-  amount: Nullable<string>;
-  timestamp: number;
-  type: VaultEventType;
-};
-
-export type SubqueryVault = {
-  id: string;
-  type: 'Type1' | 'Type2';
-  status: VaultStatus;
-  collateralAssetId: string;
-  debtAssetId: string;
-  collateralAmountReturned: string;
-};
-
-export type SubsquidVault = {
-  id: string;
-  type: 'Type1' | 'Type2';
-  status: VaultStatus;
-  collateralAsset: {
-    id: string;
-  };
-  debtAsset: {
-    id: string;
-  };
-  collateralAmountReturned: string;
-};
-
 export type ClosedVault = Omit<Vault, 'ownerId' | 'lockedAmount' | 'debt' | 'internalDebt' | 'interestCoefficient'> & {
   status: VaultStatus;
   returned: FPNumber;
